+++ wilddish_laravel/resources/views/recipes/list.blade.php
@@ -47,6 +47,11 @@
           @endforeach
           <p class="ingredients">{{$recipe->ingredients}}</p>
         </div>
+        @if(Auth::check())
+        <div class="editmove">
+          <a href="/edit/{{$recipe->id}}">編集</a>
+        </div>
+        @endif
       </div>
     @endforeach
     </div>
