+++ wilddish_laravel/routes/web.php
@@ -35,29 +35,29 @@ Route::get('/profile/myrecipes/{id}',[App\Http\Controllers\ProfileController::cl
 Route::get('/recipes/{id}', [App\Http\Controllers\RecipeController::class, 'index'])->name('recipes');
 
 // レシピ登録画面へのルート
-Route::get('/forms', [App\Http\Controllers\RecipeController::class, 'form'])->name('form');
+Route::get('/forms', [App\Http\Controllers\FormController::class, 'form'])->name('form');
 
 // レシピ一覧画面へのルート
 Route::get('/list', [App\Http\Controllers\RecipeController::class, 'list'])->name('list');
 
 // レシピ編集画面へのルート
-Route::get('/edit/{id}', [App\Http\Controllers\RecipeController::class, 'edit'])->name('edit');
+Route::get('/edit/{id}', [App\Http\Controllers\EditController::class, 'edit'])->name('edit');
 
-Route::post('/delete', [App\Http\Controllers\RecipeController::class, 'destroy'])->name('delete');
+Route::post('/delete', [App\Http\Controllers\EditController::class, 'destroy'])->name('delete');
 
-Route::post('/update/{recipeid}', [App\Http\Controllers\RecipeController::class, 'update'])->name('update');
+Route::post('/update/{recipeid}', [App\Http\Controllers\EditController::class, 'update'])->name('update');
 
-Route::post('/recipe', [App\Http\Controllers\RecipeController::class, 'process1'])->name('recipe');
+Route::post('/recipe', [App\Http\Controllers\FormController::class, 'process1'])->name('recipe');
 
-Route::post('/recipe', [App\Http\Controllers\RecipeController::class, 'process2'])->name('recipe');
+Route::post('/recipe', [App\Http\Controllers\FormController::class, 'process2'])->name('recipe');
 
-Route::post('/recipe', [App\Http\Controllers\RecipeController::class, 'process3'])->name('recipe');
+Route::post('/recipe', [App\Http\Controllers\FormController::class, 'process3'])->name('recipe');
 
-Route::post('/recipe', [App\Http\Controllers\RecipeController::class, 'process4'])->name('recipe');
+Route::post('/recipe', [App\Http\Controllers\FormController::class, 'process4'])->name('recipe');
 
-Route::post('/recipe', [App\Http\Controllers\RecipeController::class, 'process5'])->name('recipe');
+Route::post('/recipe', [App\Http\Controllers\FormController::class, 'process5'])->name('recipe');
 
-Route::post('/recipe', [App\Http\Controllers\RecipeController::class, 'store'])->name('recipe');
+Route::post('/recipe', [App\Http\Controllers\FormController::class, 'store'])->name('recipe');
 
 
 Route::get('/top', [App\Http\Controllers\topController::class, 'index'])->name('top');
